ref(settgins): Move to folder and remove logged out state handling
